refactor: better layout of visual code
