+++ frontends/web/react/components/board/board.css
@@ -1,22 +1,11 @@
 .board {
     background-color: #1b1a17;
     border: 2px solid #50cb93;
-    padding: 8px;
+    padding: 8px 8px 8px 8px;
     user-select: none;
-}
-
-.square {
-    display: inline-block;
-    cursor: pointer;
-    width: 24px;
-    height: 24px;
-    transition: transform 0.3s cubic-bezier(0.075, 0.82, 0.165, 1);
-}
-
-.square:hover {
-    transform: scale(1.4);
-}
-
-.square:active {
-    transform: scale(0.9);
+    -o-user-select: none;
+    -ms-user-select: none;
+    -moz-user-select: none;
+    -khtml-user-select: none;
+    -webkit-user-select: none;
 }
