+++ frontends/web/react/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./board/board";
+export * from "./square/square";
